requirements.txt
streamlit
crewai
numpy
pandas
requests
python-dotenv
